+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixes
+
+- Set footer as unmodifiable map
+
 ## [1.2.0] - 2022-10-30
 
 ### Features
@@ -26,6 +32,7 @@
 - Conventional commit parser
 - Conventional commit generator
 
+[Unreleased]: https://github.com/clean-code-rocks/conventional-commit/compare/v1.2.0...main
 [1.2.0]: https://github.com/clean-code-rocks/conventional-commit/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/clean-code-rocks/conventional-commit/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/clean-code-rocks/conventional-commit/releases/tag/v1.0.0